@@ -1,23 +1,10 @@
 <Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" Exists('$(MSBuildProjectDirectory)\..\..\..\src\Common.CoreLib\Properties\ThisAssembly.Version.cs') ">
-    <DefineConstants>UnifiedVersion;$(DefineConstants)</DefineConstants>
-  </PropertyGroup>
-
-  <!--<ItemGroup>
-    <PackageReference Include="Microsoft.Win32.Registry" />
-  </ItemGroup>-->
-
-  <ItemGroup Condition=" Exists('$(MSBuildProjectDirectory)\..\..\..\src\Common.CoreLib\Properties\ThisAssembly.Version.cs') ">
-    <Compile Include="..\..\..\src\Common.CoreLib\Properties\AssemblyInfo.Version.cs">
-      <Link>Properties\AssemblyInfo.Version.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\src\Common.CoreLib\Properties\ThisAssembly.Version.cs">
-      <Link>Properties\ThisAssembly.Version.cs</Link>
-    </Compile>
-  </ItemGroup>
+	<PropertyGroup>
+		<TargetFramework>net7.0</TargetFramework>
+		<LangVersion>latest</LangVersion>
+		<Nullable>enable</Nullable>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<IsTrimmable>true</IsTrimmable>
+		<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+	</PropertyGroup>
 </Project>
