@@ -20,188 +20,182 @@
  *    distribution.
  */
 
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.Globalization;
-using System.IO;
-using System.Linq;
 
-namespace SAM.API
+namespace SAM.API;
+
+public class Client : IDisposable
 {
-    public class Client : IDisposable
+    public Wrappers.SteamClient018 SteamClient { get; set; }
+    public Wrappers.SteamUser017 SteamUser { get; set; }
+    public Wrappers.SteamUserStats011 SteamUserStats { get; set; }
+    public Wrappers.SteamUtils007 SteamUtils { get; set; }
+    public Wrappers.SteamApps001 SteamApps001 { get; set; }
+    public Wrappers.SteamApps008 SteamApps008 { get; set; }
+    public Wrappers.SteamRemoteStorage012 SteamRemoteStorage { get; set; }
+    public bool IsConnectToSteam { get; set; }
+
+    private bool _IsDisposed = false;
+    private int _Pipe;
+    private int _User;
+
+    private readonly List<ICallback> _Callbacks = new();
+
+    const string KEY_STEAM_APP_ID = "SteamAppId";
+
+    public static bool WriteSteamAppIdTxt { get; set; }
+
+    public bool Initialize(long appId)
     {
-        public Wrappers.SteamClient018 SteamClient { get; set; }
-        public Wrappers.SteamUser017 SteamUser { get; set; }
-        public Wrappers.SteamUserStats011 SteamUserStats { get; set; }
-        public Wrappers.SteamUtils007 SteamUtils { get; set; }
-        public Wrappers.SteamApps001 SteamApps001 { get; set; }
-        public Wrappers.SteamApps008 SteamApps008 { get; set; }
-        public Wrappers.SteamRemoteStorage012 SteamRemoteStorage { get; set; }
-        public bool IsConnectToSteam { get; set; }
-
-        private bool _IsDisposed = false;
-        private int _Pipe;
-        private int _User;
-
-        private readonly List<ICallback> _Callbacks = new();
-
-        const string KEY_STEAM_APP_ID = "SteamAppId";
-
-        public static bool WriteSteamAppIdTxt { get; set; }
-
-        public bool Initialize(long appId)
+        if (string.IsNullOrEmpty(Steam.GetInstallPath()) == true)
         {
-            if (string.IsNullOrEmpty(Steam.GetInstallPath()) == true)
+            throw new ClientInitializeException(ClientInitializeFailure.GetInstallPath, "failed to get Steam install path");
+        }
+
+        string steam_appid_file_path = null;
+
+        try
+        {
+            if (appId != 0)
             {
-                throw new ClientInitializeException(ClientInitializeFailure.GetInstallPath, "failed to get Steam install path");
-            }
-
-            string steam_appid_file_path = null;
-
-            try
-            {
-                if (appId != 0)
+                Environment.SetEnvironmentVariable(KEY_STEAM_APP_ID, appId.ToString(CultureInfo.InvariantCulture));
+                
+                if (Steam.Load() == false)
                 {
-                    Environment.SetEnvironmentVariable(KEY_STEAM_APP_ID, appId.ToString(CultureInfo.InvariantCulture));
-                    
-                    if (Steam.Load() == false)
-                    {
-                        steam_appid_file_path = Path.Combine(Path.GetDirectoryName(Environment.ProcessPath), "steam_appid.txt");
-                        File.WriteAllText(steam_appid_file_path, appId.ToString());
+                    steam_appid_file_path = Path.Combine(Path.GetDirectoryName(Environment.ProcessPath), "steam_appid.txt");
+                    File.WriteAllText(steam_appid_file_path, appId.ToString());
 
-                        if (Steam.Load() == false)
-                        {
-                            throw new ClientInitializeException(ClientInitializeFailure.Load, "failed to load Steam");
-                        }
-                    }
-                }
-                else
-                {
                     if (Steam.Load() == false)
                     {
                         throw new ClientInitializeException(ClientInitializeFailure.Load, "failed to load Steam");
                     }
                 }
-
-                this.SteamClient = Steam.CreateInterface<Wrappers.SteamClient018>("SteamClient018");
-                if (this.SteamClient == null)
-                {
-                    throw new ClientInitializeException(ClientInitializeFailure.CreateSteamClient, "failed to create ISteamClient017");
-                }
-
-                this._Pipe = this.SteamClient.CreateSteamPipe();
-                if (this._Pipe == 0)
-                {
-                    throw new ClientInitializeException(ClientInitializeFailure.CreateSteamPipe, "failed to create pipe");
-                }
-
-                this._User = this.SteamClient.ConnectToGlobalUser(this._Pipe);
-                if (this._User == 0)
-                {
-                    throw new ClientInitializeException(ClientInitializeFailure.ConnectToGlobalUser, "failed to connect to global user");
-                }
-
-                this.SteamUtils = this.SteamClient.GetSteamUtils007(this._Pipe);
-                var currentAppId = this.SteamUtils.GetAppId();
-                if (appId > 0 && currentAppId != (uint)appId)
-                {
-                    //var envAppId = Environment.GetEnvironmentVariable(KEY_STEAM_APP_ID);
-                    throw new ClientInitializeException(ClientInitializeFailure.AppIdMismatch, $"appID mismatch, appId: {appId}"); //, currentAppId: {currentAppId}, envAppId: {envAppId}");
-                }
-
-                this.SteamUser = this.SteamClient.GetSteamUser017(this._User, this._Pipe);
-                this.SteamUserStats = this.SteamClient.GetSteamUserStats011(this._User, this._Pipe);
-                this.SteamApps001 = this.SteamClient.GetSteamApps001(this._User, this._Pipe);
-                this.SteamApps008 = this.SteamClient.GetSteamApps008(this._User, this._Pipe);
-                this.SteamRemoteStorage = this.SteamClient.GetSteamRemoteStorage012(this._User, this._Pipe);
             }
-            finally
+            else
             {
-                if (steam_appid_file_path != null)
+                if (Steam.Load() == false)
                 {
-                    try
-                    {
-                        File.Delete(steam_appid_file_path);
-                    }
-                    catch
-                    {
-
-                    }
+                    throw new ClientInitializeException(ClientInitializeFailure.Load, "failed to load Steam");
                 }
             }
 
-            return true;
-        }
-
-        ~Client()
-        {
-            this.Dispose(false);
-        }
-
-        protected virtual void Dispose(bool disposing)
-        {
-            if (this._IsDisposed == true)
+            SteamClient = Steam.CreateInterface<Wrappers.SteamClient018>("SteamClient018");
+            if (SteamClient == null)
             {
-                return;
+                throw new ClientInitializeException(ClientInitializeFailure.CreateSteamClient, "failed to create ISteamClient017");
             }
 
-            this._IsDisposed = true;
-
-            if (this.SteamClient != null && this._Pipe > 0)
+            _Pipe = SteamClient.CreateSteamPipe();
+            if (_Pipe == 0)
             {
-                if (this._User > 0)
+                throw new ClientInitializeException(ClientInitializeFailure.CreateSteamPipe, "failed to create pipe");
+            }
+
+            _User = SteamClient.ConnectToGlobalUser(_Pipe);
+            if (_User == 0)
+            {
+                throw new ClientInitializeException(ClientInitializeFailure.ConnectToGlobalUser, "failed to connect to global user");
+            }
+
+            SteamUtils = SteamClient.GetSteamUtils007(_Pipe);
+            var currentAppId = SteamUtils.GetAppId();
+            if (appId > 0 && currentAppId != (uint)appId)
+            {
+                //var envAppId = Environment.GetEnvironmentVariable(KEY_STEAM_APP_ID);
+                throw new ClientInitializeException(ClientInitializeFailure.AppIdMismatch, $"appID mismatch, appId: {appId}"); //, currentAppId: {currentAppId}, envAppId: {envAppId}");
+            }
+
+            SteamUser = SteamClient.GetSteamUser017(_User, _Pipe);
+            SteamUserStats = SteamClient.GetSteamUserStats011(_User, _Pipe);
+            SteamApps001 = SteamClient.GetSteamApps001(_User, _Pipe);
+            SteamApps008 = SteamClient.GetSteamApps008(_User, _Pipe);
+            SteamRemoteStorage = SteamClient.GetSteamRemoteStorage012(_User, _Pipe);
+        }
+        finally
+        {
+            if (steam_appid_file_path != null)
+            {
+                try
                 {
-                    this.SteamClient.ReleaseUser(this._Pipe, this._User);
-                    this._User = 0;
+                    File.Delete(steam_appid_file_path);
                 }
-
-                this.SteamClient.ReleaseSteamPipe(this._Pipe);
-                this._Pipe = 0;
-            }
-
-            this._IsDisposed = false;
-        }
-
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        public TCallback CreateAndRegisterCallback<TCallback>()
-            where TCallback : ICallback, new()
-        {
-            var callback = new TCallback();
-            this._Callbacks.Add(callback);
-            return callback;
-        }
-
-        private bool _RunningCallbacks;
-
-        public void RunCallbacks(bool server)
-        {
-            if (this._RunningCallbacks == true)
-            {
-                return;
-            }
-
-            this._RunningCallbacks = true;
-
-            while (Steam.GetCallback(this._Pipe, out Types.CallbackMessage message, out int call) == true)
-            {
-                var callbackId = message.Id;
-                foreach (ICallback callback in this._Callbacks.Where(
-                    candidate => candidate.Id == callbackId &&
-                                 candidate.IsServer == server))
+                catch
                 {
-                    callback.Run(message.ParamPointer);
-                }
 
-                Steam.FreeLastCallback(this._Pipe);
+                }
+            }
+        }
+
+        return true;
+    }
+
+    ~Client()
+    {
+        Dispose(false);
+    }
+
+    protected virtual void Dispose(bool disposing)
+    {
+        if (_IsDisposed == true)
+        {
+            return;
+        }
+
+        _IsDisposed = true;
+
+        if (SteamClient != null && _Pipe > 0)
+        {
+            if (_User > 0)
+            {
+                SteamClient.ReleaseUser(_Pipe, _User);
+                _User = 0;
             }
 
-            this._RunningCallbacks = false;
+            SteamClient.ReleaseSteamPipe(_Pipe);
+            _Pipe = 0;
         }
+
+        _IsDisposed = false;
+    }
+
+    public void Dispose()
+    {
+        Dispose(true);
+        GC.SuppressFinalize(this);
+    }
+
+    public TCallback CreateAndRegisterCallback<TCallback>()
+        where TCallback : ICallback, new()
+    {
+        var callback = new TCallback();
+        _Callbacks.Add(callback);
+        return callback;
+    }
+
+    private bool _RunningCallbacks;
+
+    public void RunCallbacks(bool server)
+    {
+        if (_RunningCallbacks == true)
+        {
+            return;
+        }
+
+        _RunningCallbacks = true;
+
+        while (Steam.GetCallback(_Pipe, out Types.CallbackMessage message, out int call) == true)
+        {
+            var callbackId = message.Id;
+            foreach (ICallback callback in _Callbacks.Where(
+                candidate => candidate.Id == callbackId &&
+                             candidate.IsServer == server))
+            {
+                callback.Run(message.ParamPointer);
+            }
+
+            Steam.FreeLastCallback(_Pipe);
+        }
+
+        _RunningCallbacks = false;
     }
 }
