@@ -19,32 +19,30 @@
  * 3. This notice may not be removed or altered from any source
  *    distribution.
  */
-using System;
 
-namespace SAM.API
+namespace SAM.API;
+
+public class ClientInitializeException : Exception
 {
-    public class ClientInitializeException : Exception
+    public readonly ClientInitializeFailure Failure;
+
+    public ClientInitializeException(ClientInitializeFailure failure)
     {
-        public readonly ClientInitializeFailure Failure;
+        Failure = failure;
+    }
 
-        public ClientInitializeException(ClientInitializeFailure failure)
-        {
-            this.Failure = failure;
-        }
+    public ClientInitializeException(ClientInitializeFailure failure, string message)
+        : base(message)
+    {
+        Failure = failure;
+    }
 
-        public ClientInitializeException(ClientInitializeFailure failure, string message)
-            : base(message)
-        {
-            this.Failure = failure;
-        }
-
-        public ClientInitializeException(
-            ClientInitializeFailure failure,
-            string message,
-            Exception innerException)
-            : base(message, innerException)
-        {
-            this.Failure = failure;
-        }
+    public ClientInitializeException(
+        ClientInitializeFailure failure,
+        string message,
+        Exception innerException)
+        : base(message, innerException)
+    {
+        Failure = failure;
     }
 }
