@@ -20,14 +20,12 @@
  *    distribution.
  */
 
-using System;
 using System.Runtime.InteropServices;
 
-namespace SAM.API
+namespace SAM.API;
+
+[StructLayout(LayoutKind.Sequential, Pack = 1, CharSet = CharSet.Ansi)]
+internal struct NativeClass
 {
-    [StructLayout(LayoutKind.Sequential, Pack = 1, CharSet = CharSet.Ansi)]
-    internal struct NativeClass
-    {
-        public IntPtr VirtualTable;
-    }
+    public IntPtr VirtualTable;
 }
