@@ -20,86 +20,80 @@
  *    distribution.
  */
 
-using System;
-using System.Collections.Generic;
 using System.Runtime.InteropServices;
-using System.Text;
 
-namespace SAM.API
+namespace SAM.API;
+
+public abstract class NativeWrapper<TNativeFunctions> : INativeWrapper
 {
-    public abstract class NativeWrapper<TNativeFunctions> : INativeWrapper
+    protected IntPtr ObjectAddress;
+    protected TNativeFunctions? Functions;
+    public override string ToString()
     {
-        protected IntPtr ObjectAddress;
-        protected TNativeFunctions Functions;
-        public override string ToString()
+        return string.Format(
+            System.Globalization.CultureInfo.CurrentCulture,
+            "Steam Interface<{0}> #{1:X8}",
+            typeof(TNativeFunctions),
+            ObjectAddress.ToInt32());
+    }
+
+    public void SetupFunctions(IntPtr objectAddress)
+    {
+        ObjectAddress = objectAddress;
+
+        var iface = Marshal.PtrToStructure<NativeClass>(
+            ObjectAddress);
+
+        Functions = Marshal.PtrToStructure<TNativeFunctions>(
+            iface.VirtualTable);
+    }
+
+    private readonly Dictionary<IntPtr, Delegate> _FunctionCache = new();
+
+    protected Delegate GetDelegate<TDelegate>(IntPtr pointer)
+    {
+        Delegate function;
+
+        if (_FunctionCache.ContainsKey(pointer) == false)
         {
-            return string.Format(
-                System.Globalization.CultureInfo.CurrentCulture,
-                "Steam Interface<{0}> #{1:X8}",
-                typeof(TNativeFunctions),
-                this.ObjectAddress.ToInt32());
+            function = Marshal.GetDelegateForFunctionPointer(pointer, typeof(TDelegate));
+            _FunctionCache[pointer] = function;
+        }
+        else
+        {
+            function = _FunctionCache[pointer];
         }
 
-        public void SetupFunctions(IntPtr objectAddress)
+        return function;
+    }
+
+    protected TDelegate GetFunction<TDelegate>(IntPtr pointer)
+        where TDelegate : class
+    {
+        return (TDelegate)((object)GetDelegate<TDelegate>(pointer));
+    }
+
+    protected void Call<TDelegate>(IntPtr pointer, params object[] args)
+    {
+        try
         {
-            this.ObjectAddress = objectAddress;
-
-            var iface = (NativeClass)Marshal.PtrToStructure(
-                this.ObjectAddress,
-                typeof(NativeClass));
-
-            this.Functions = (TNativeFunctions)Marshal.PtrToStructure(
-                iface.VirtualTable,
-                typeof(TNativeFunctions));
+            GetDelegate<TDelegate>(pointer).DynamicInvoke(args);
         }
-
-        private readonly Dictionary<IntPtr, Delegate> _FunctionCache = new Dictionary<IntPtr, Delegate>();
-
-        protected Delegate GetDelegate<TDelegate>(IntPtr pointer)
+        catch (Exception)
         {
-            Delegate function;
-
-            if (this._FunctionCache.ContainsKey(pointer) == false)
-            {
-                function = Marshal.GetDelegateForFunctionPointer(pointer, typeof(TDelegate));
-                this._FunctionCache[pointer] = function;
-            }
-            else
-            {
-                function = this._FunctionCache[pointer];
-            }
-
-            return function;
+            throw;
         }
+    }
 
-        protected TDelegate GetFunction<TDelegate>(IntPtr pointer)
-            where TDelegate : class
+    protected TReturn? Call<TReturn, TDelegate>(IntPtr pointer, params object[] args)
+    {
+        try
         {
-            return (TDelegate)((object)this.GetDelegate<TDelegate>(pointer));
+            return (TReturn?)GetDelegate<TDelegate>(pointer).DynamicInvoke(args);
         }
-
-        protected void Call<TDelegate>(IntPtr pointer, params object[] args)
+        catch (Exception)
         {
-            try
-            {
-                this.GetDelegate<TDelegate>(pointer).DynamicInvoke(args);
-            }
-            catch (Exception)
-            {
-                throw;
-            }
-        }
-
-        protected TReturn? Call<TReturn, TDelegate>(IntPtr pointer, params object[] args)
-        {
-            try
-            {
-                return (TReturn?)this.GetDelegate<TDelegate>(pointer).DynamicInvoke(args);
-            }
-            catch (Exception)
-            {
-                throw;
-            }
+            throw;
         }
     }
 }
