@@ -20,19 +20,18 @@
  *    distribution.
  */
 
-namespace SAM.API.Types
+namespace SAM.API.Types;
+
+public enum AccountType : int
 {
-    public enum AccountType : int
-    {
-        Invalid = 0,
-        Individual = 1,
-        Multiset = 2,
-        GameServer = 3,
-        AnonGameServer = 4,
-        Pending = 5,
-        ContentServer = 6,
-        Clan = 7,
-        Chat = 8,
-        P2PSuperSeeder = 9,
-    }
+    Invalid = 0,
+    Individual = 1,
+    Multiset = 2,
+    GameServer = 3,
+    AnonGameServer = 4,
+    Pending = 5,
+    ContentServer = 6,
+    Clan = 7,
+    Chat = 8,
+    P2PSuperSeeder = 9,
 }
