@@ -22,12 +22,11 @@
 
 using System.Runtime.InteropServices;
 
-namespace SAM.API.Types
+namespace SAM.API.Types;
+
+[StructLayout(LayoutKind.Sequential, Pack = 1)]
+public struct AppDataChanged
 {
-    [StructLayout(LayoutKind.Sequential, Pack = 1)]
-    public struct AppDataChanged
-    {
-        public uint Id;
-        public bool Result;
-    }
+    public uint Id;
+    public bool Result;
 }
