@@ -20,18 +20,17 @@
  *    distribution.
  */
 
-namespace SAM.API.Types
+namespace SAM.API.Types;
+
+public enum ItemRequestResult : int
 {
-    public enum ItemRequestResult : int
-    {
-        InvalidValue = -1,
-        OK = 0,
-        Denied = 1,
-        ServerError = 2,
-        Timeout = 3,
-        Invalid = 4,
-        NoMatch = 5,
-        UnknownError = 6,
-        NotLoggedOn = 7,
-    }
+    InvalidValue = -1,
+    OK = 0,
+    Denied = 1,
+    ServerError = 2,
+    Timeout = 3,
+    Invalid = 4,
+    NoMatch = 5,
+    UnknownError = 6,
+    NotLoggedOn = 7,
 }
