@@ -20,15 +20,14 @@
  *    distribution.
  */
 
-namespace SAM.API.Types
+namespace SAM.API.Types;
+
+public enum UserStatType
 {
-    public enum UserStatType
-    {
-        Invalid = 0,
-        Integer = 1,
-        Float = 2,
-        AverageRate = 3,
-        Achievements = 4,
-        GroupAchievements = 5,
-    }
+    Invalid = 0,
+    Integer = 1,
+    Float = 2,
+    AverageRate = 3,
+    Achievements = 4,
+    GroupAchievements = 5,
 }
