@@ -22,12 +22,11 @@
 
 using System.Runtime.InteropServices;
 
-namespace SAM.API.Types
+namespace SAM.API.Types;
+
+[StructLayout(LayoutKind.Sequential, Pack = 1)]
+public struct UserStatsStored
 {
-    [StructLayout(LayoutKind.Sequential, Pack = 1)]
-    public struct UserStatsStored
-    {
-        public ulong GameId;
-        public int Result;
-    }
+    public ulong GameId;
+    public int Result;
 }
