@@ -21,102 +21,100 @@
  */
 
 using SAM.API.Interfaces;
-using System;
 using System.Runtime.InteropServices;
 
-namespace SAM.API.Wrappers
+namespace SAM.API.Wrappers;
+
+public class SteamApps008 : NativeWrapper<ISteamApps008>
 {
-    public class SteamApps008 : NativeWrapper<ISteamApps008>
+    #region IsSubscribed
+    [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
+    [return: MarshalAs(UnmanagedType.I1)]
+    private delegate bool NativeIsSubscribedApp(IntPtr self, uint appID);
+
+    public bool IsSubscribedApp(uint appID)
     {
-        #region IsSubscribed
-        [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
-        [return: MarshalAs(UnmanagedType.I1)]
-        private delegate bool NativeIsSubscribedApp(IntPtr self, uint appID);
-
-        public bool IsSubscribedApp(uint appID)
-        {
-            return this.Call<bool, NativeIsSubscribedApp>(this.Functions.IsSubscribedApp, this.ObjectAddress, appID);
-        }
-        #endregion
-
-        #region GetCurrentGameLanguage
-        [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
-        private delegate IntPtr NativeGetCurrentGameLanguage(IntPtr self);
-
-        public string GetCurrentGameLanguage()
-        {
-            var languagePointer = this.Call<IntPtr, NativeGetCurrentGameLanguage>(
-                this.Functions.GetCurrentGameLanguage,
-                this.ObjectAddress);
-            return NativeStrings.PointerToString(languagePointer);
-        }
-        #endregion
-
-        #region GetAvailableGameLanguages
-        [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
-        private delegate IntPtr NativeGetAvailableGameLanguages(IntPtr self);
-
-        public string GetAvailableGameLanguages()
-        {
-            var languagePointer = this.Call<IntPtr, NativeGetAvailableGameLanguages>(
-                this.Functions.GetAvailableGameLanguages,
-                this.ObjectAddress);
-            return NativeStrings.PointerToString(languagePointer);
-        }
-
-        #endregion
-
-        #region IsDlcInstalled
-        [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
-        [return: MarshalAs(UnmanagedType.I1)]
-        private delegate bool NativeIsDlcInstalled(IntPtr self, uint appid);
-        public bool IsDlcInstalled(uint appid)
-        {
-            return this.Call<bool, NativeIsDlcInstalled>(this.Functions.IsDlcInstalled, this.ObjectAddress, appid);
-        }
-        #endregion
-
-        #region IsAppInstalled
-        [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
-        [return: MarshalAs(UnmanagedType.I1)]
-        private delegate bool NativeIsAppInstalled(IntPtr self, uint appid);
-        public bool IsAppInstalled(uint appid)
-        {
-            return this.Call<bool, NativeIsAppInstalled>(this.Functions.IsAppInstalled, this.ObjectAddress, appid);
-        }
-        #endregion
-
-        #region IsSubscribedFromFamilySharing
-        [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
-        [return: MarshalAs(UnmanagedType.I1)]
-        private delegate bool NativeIsSubscribedFromFamilySharing(IntPtr self);
-        public bool IsSubscribedFromFamilySharing()
-        {
-            return this.Call<bool, NativeIsSubscribedFromFamilySharing>(this.Functions.IsSubscribedFromFamilySharing, this.ObjectAddress);
-        }
-        #endregion
-
-        #region IsSubscribedFromFreeWeekend
-        [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
-        [return: MarshalAs(UnmanagedType.I1)]
-        private delegate bool NativeIsSubscribedFromFreeWeekend(IntPtr self);
-        public bool IsSubscribedFromFreeWeekend()
-        {
-            return this.Call<bool, NativeIsSubscribedFromFreeWeekend>(this.Functions.IsSubscribedFromFreeWeekend, this.ObjectAddress);
-        }
-        #endregion
-
-        #region GetAppInstallDir
-        [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
-        private delegate IntPtr NativeGetAppInstallDir(IntPtr self, uint appid, IntPtr pchFolder, uint cchFolderBufferSize);
-        public string GetAppInstallDir(uint appID)
-        {
-            IntPtr mempchFolder = Helpers.TakeMemory();
-            string pchFolder;
-            this.GetFunction<NativeGetAppInstallDir>(this.Functions.GetAppInstallDir)(this.ObjectAddress, appID, mempchFolder, (1024 * 32));
-            pchFolder = Helpers.MemoryToString(mempchFolder);
-            return pchFolder;
-        }
-        #endregion
+        return Call<bool, NativeIsSubscribedApp>(Functions.IsSubscribedApp, ObjectAddress, appID);
     }
+    #endregion
+
+    #region GetCurrentGameLanguage
+    [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
+    private delegate IntPtr NativeGetCurrentGameLanguage(IntPtr self);
+
+    public string GetCurrentGameLanguage()
+    {
+        var languagePointer = Call<IntPtr, NativeGetCurrentGameLanguage>(
+            Functions.GetCurrentGameLanguage,
+            ObjectAddress);
+        return NativeStrings.PointerToString(languagePointer);
+    }
+    #endregion
+
+    #region GetAvailableGameLanguages
+    [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
+    private delegate IntPtr NativeGetAvailableGameLanguages(IntPtr self);
+
+    public string GetAvailableGameLanguages()
+    {
+        var languagePointer = Call<IntPtr, NativeGetAvailableGameLanguages>(
+            Functions.GetAvailableGameLanguages,
+            ObjectAddress);
+        return NativeStrings.PointerToString(languagePointer);
+    }
+
+    #endregion
+
+    #region IsDlcInstalled
+    [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
+    [return: MarshalAs(UnmanagedType.I1)]
+    private delegate bool NativeIsDlcInstalled(IntPtr self, uint appid);
+    public bool IsDlcInstalled(uint appid)
+    {
+        return Call<bool, NativeIsDlcInstalled>(Functions.IsDlcInstalled, ObjectAddress, appid);
+    }
+    #endregion
+
+    #region IsAppInstalled
+    [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
+    [return: MarshalAs(UnmanagedType.I1)]
+    private delegate bool NativeIsAppInstalled(IntPtr self, uint appid);
+    public bool IsAppInstalled(uint appid)
+    {
+        return Call<bool, NativeIsAppInstalled>(Functions.IsAppInstalled, ObjectAddress, appid);
+    }
+    #endregion
+
+    #region IsSubscribedFromFamilySharing
+    [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
+    [return: MarshalAs(UnmanagedType.I1)]
+    private delegate bool NativeIsSubscribedFromFamilySharing(IntPtr self);
+    public bool IsSubscribedFromFamilySharing()
+    {
+        return Call<bool, NativeIsSubscribedFromFamilySharing>(Functions.IsSubscribedFromFamilySharing, ObjectAddress);
+    }
+    #endregion
+
+    #region IsSubscribedFromFreeWeekend
+    [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
+    [return: MarshalAs(UnmanagedType.I1)]
+    private delegate bool NativeIsSubscribedFromFreeWeekend(IntPtr self);
+    public bool IsSubscribedFromFreeWeekend()
+    {
+        return Call<bool, NativeIsSubscribedFromFreeWeekend>(Functions.IsSubscribedFromFreeWeekend, ObjectAddress);
+    }
+    #endregion
+
+    #region GetAppInstallDir
+    [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
+    private delegate IntPtr NativeGetAppInstallDir(IntPtr self, uint appid, IntPtr pchFolder, uint cchFolderBufferSize);
+    public string GetAppInstallDir(uint appID)
+    {
+        IntPtr mempchFolder = Helpers.TakeMemory();
+        string pchFolder;
+        GetFunction<NativeGetAppInstallDir>(Functions.GetAppInstallDir)(ObjectAddress, appID, mempchFolder, (1024 * 32));
+        pchFolder = Helpers.MemoryToString(mempchFolder);
+        return pchFolder;
+    }
+    #endregion
 }
