@@ -21,35 +21,33 @@
  */
 
 using SAM.API.Interfaces;
-using System;
 using System.Runtime.InteropServices;
 
-namespace SAM.API.Wrappers
+namespace SAM.API.Wrappers;
+
+public class SteamUser012 : NativeWrapper<ISteamUser012>
 {
-    public class SteamUser012 : NativeWrapper<ISteamUser012>
+    #region IsLoggedIn
+    [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
+    [return: MarshalAs(UnmanagedType.I1)]
+    private delegate bool NativeLoggedOn(IntPtr self);
+
+    public bool IsLoggedIn()
     {
-        #region IsLoggedIn
-        [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
-        [return: MarshalAs(UnmanagedType.I1)]
-        private delegate bool NativeLoggedOn(IntPtr self);
-
-        public bool IsLoggedIn()
-        {
-            return this.Call<bool, NativeLoggedOn>(this.Functions.LoggedOn, this.ObjectAddress);
-        }
-        #endregion
-
-        #region GetSteamID
-        [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
-        private delegate void NativeGetSteamId(IntPtr self, out ulong steamId);
-
-        public ulong GetSteamId()
-        {
-            var call = this.GetFunction<NativeGetSteamId>(this.Functions.GetSteamID);
-            ulong steamId;
-            call(this.ObjectAddress, out steamId);
-            return steamId;
-        }
-        #endregion
+        return Call<bool, NativeLoggedOn>(Functions.LoggedOn, ObjectAddress);
     }
+    #endregion
+
+    #region GetSteamID
+    [UnmanagedFunctionPointer(CallingConvention.ThisCall)]
+    private delegate void NativeGetSteamId(IntPtr self, out ulong steamId);
+
+    public ulong GetSteamId()
+    {
+        var call = GetFunction<NativeGetSteamId>(Functions.GetSteamID);
+        ulong steamId;
+        call(ObjectAddress, out steamId);
+        return steamId;
+    }
+    #endregion
 }
