@@ -20,14 +20,11 @@
  *    distribution.
  */
 
-using System;
+namespace SAM.API;
 
-namespace SAM.API
+public interface ICallback
 {
-    public interface ICallback
-    {
-        int Id { get; }
-        bool IsServer { get; }
-        void Run(IntPtr param);
-    }
+    int Id { get; }
+    bool IsServer { get; }
+    void Run(IntPtr param);
 }
